requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
pandas
numpy
python-dotenv